+++ core/test/unit/middleware_spec.js
@@ -29,74 +29,6 @@ describe('Middleware', function () {
     afterEach(function () {
         sandbox.restore();
     });
-    // TODO: needs new test for ember admin
-    // describe('redirectToDashboard', function () {
-    //     var req, res;
-
-    //     beforeEach(function () {
-    //         req = {
-    //             session: {}
-    //         };
-
-    //         res = {
-    //             redirect: sinon.spy()
-    //         };
-    //     });
-
-    //     it('should redirect to dashboard', function () {
-    //         req.session.user = {};
-
-    //         middleware.redirectToDashboard(req, res, null);
-    //         assert(res.redirect.calledWithMatch('/ghost/'));
-    //     });
-
-    //     it('should call next if no user in session', function (done) {
-    //         middleware.redirectToDashboard(req, res, function (a) {
-    //             should.not.exist(a);
-    //             assert(res.redirect.calledOnce.should.be.false);
-    //             done();
-    //         });
-    //     });
-    // });
-
-    describe('cacheControl', function () {
-        var res;
-
-        beforeEach(function () {
-            res = {
-                set: sinon.spy()
-            };
-        });
-
-        it('correctly sets the public profile headers', function (done) {
-            middleware.cacheControl('public')(null, res, function (a) {
-                should.not.exist(a);
-                res.set.calledOnce.should.be.true;
-                res.set.calledWith({'Cache-Control': 'public, max-age=0'});
-                done();
-            });
-        });
-
-        it('correctly sets the private profile headers', function (done) {
-            middleware.cacheControl('private')(null, res, function (a) {
-                should.not.exist(a);
-                res.set.calledOnce.should.be.true;
-                res.set.calledWith({
-                    'Cache-Control':
-                        'no-cache, private, no-store, must-revalidate, max-stale=0, post-check=0, pre-check=0'
-                });
-                done();
-            });
-        });
-
-        it('will not set headers without a profile', function (done) {
-            middleware.cacheControl()(null, res, function (a) {
-                should.not.exist(a);
-                res.set.called.should.be.false;
-                done();
-            });
-        });
-    });
 
     describe('whenEnabled', function () {
         var cbFn, blogApp;
